LayoutTests/platform/ios-sim/media/csp-blocks-video-expected.txt

CONSOLE MESSAGE: Refused to load media from '' because of Content-Security-Policy.

CONSOLE MESSAGE: Refused to load media from '' because of Content-Security-Policy.

CONSOLE MESSAGE: Refused to load media from 'test.mp4' because of Content-Security-Policy.

CONSOLE MESSAGE: Refused to load media from '' because of Content-Security-Policy.

END OF TEST
This test passes if it doesn't alert failure.
